chore: extract devnet-sdk package
This will host the code that interacts with:
- the definition of a devnet (including generation of kurtosis inputs)
- an arbitrary devnet, regardless of whether it's kurtosis-backed or not

On the other hand, kurtosis-devnet remains focused solely on the
instantiation of a devnet on kurtosis.

As such, kurtosis-devnet can depend on devnet-sdk, but not the other
way around.
