+++ devnet-sdk/manifest/acceptor.go
