+++ kurtosis-devnet/pkg/kurtosis/endpoints_test.go
@@ -3,7 +3,7 @@ package kurtosis
 import (
 	"testing"
 
-	"github.com/ethereum-optimism/optimism/kurtosis-devnet/pkg/descriptors"
+	"github.com/ethereum-optimism/optimism/devnet-sdk/descriptors"
 	"github.com/ethereum-optimism/optimism/kurtosis-devnet/pkg/kurtosis/sources/inspect"
 	"github.com/stretchr/testify/assert"
 )
